requirements.txt
# Core ML Libraries
numpy
pandas
scikit-learn
xgboost
tensorflow

# Web Scraping & API
requests
cloudscraper
beautifulsoup4
fake-useragent

# Web Application
flask
flask-cors

# Visualization
matplotlib
seaborn

# Utilities
tqdm
joblib

# Optional but recommended
jupyter
notebook
python-dotenv

# # Core ML Libraries
# numpy==1.24.3
# pandas==2.0.3
# scikit-learn==1.3.0
# xgboost==2.0.0
# tensorflow==2.13.0

# # Web Scraping & API
# requests==2.31.0
# cloudscraper==1.2.71
# beautifulsoup4==4.12.2
# fake-useragent==1.3.0

# # Web Application
# flask==2.3.3
# flask-cors==4.0.0

# # Visualization
# matplotlib==3.7.2
# seaborn==0.12.2

# # Utilities
# tqdm==4.66.1
# joblib==1.3.2

# # Optional but recommended
# jupyter==1.0.0
# notebook==7.0.3